Break apart the scan scene styles
