+++ src/constants/IconConstants.js
@@ -39,7 +39,6 @@ export const MAIL = 'ios-mail'
 export const ALERT_OUTLINE = 'ios-alert-outline'
 export const KEY_ICON = 'ios-key'
 export const WALLET_ICON = 'ios-wallet'
-export const WARNING = 'ios-warning'
 
 // Material
 export const MATERIAL_ICONS = 'material'
